pyproject.toml
[project]
name = "base-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "cookiecutter>=2.6.0",
    "django>=5.2.5",
    "django-tailwind[reload]>=4.2.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "honcho>=2.0.0",
    "mysqlclient>=2.2.7",
]
